backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
asyncpg>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
aiosqlite>=0.19.0
python-multipart>=0.0.12
httpx>=0.28.0
openai>=1.50.0
google-generativeai>=0.8.0
python-docx>=1.1.0
PyPDF2>=3.0.0
alembic>=1.14.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
